requirements.txt
anthropic
requests
PyYAML
